+++ tests/functional/navigator/test_navigator.py
@@ -131,17 +131,6 @@ class TestNavigator(TestCase):
                                                        "About",
                                                        screen_change_before_first_instruction=False)
 
-    def test_navigate_until_text_cannot_find_text(self):
-        with patch("speculos.client.subprocess"):
-            with SpeculosServerStub():
-                with self.backend:
-                    with self.assertRaises(TimeoutError) as error:
-                        self.navigator.navigate_until_text(
-                            NavIns(NavInsID.RIGHT_CLICK), [NavIns(NavInsID.BOTH_CLICK)],
-                            "WILL NOT BE FOUND",
-                            screen_change_before_first_instruction=False)
-                    self.assertIn("Timeout waiting for text", str(error.exception))
-
     def test_navigate_until_text_screen_change_timeout(self):
         with patch("speculos.client.subprocess"):
             with SpeculosServerStub():